python_renderer/requirements.txt
colormath==3.0.0
networkx==3.4.2
numpy==2.2.6
opencv-python==4.13.0.92
posix_ipc==1.3.2
pygame==2.6.1
PyOpenGL==3.1.10
python-json-logger==4.0.0
